// types/dog.tsx
// types/index.tsx

export interface Dog {
    id: string;
    description: string;
    temperament: string;
    popularity: number;
    min_height: number;
    max_height: number;
    min_weight: number;
    max_weight: number;
    min_expectancy: number;
    max_expectancy: number;
    group: string;
    grooming_frequency_value: number;
    grooming_frequency_category: string;
    shedding_value: number;
    shedding_category: string;
    energy_level_value: number;
    energy_level_category: string;
    trainability_value: number;
    trainability_category: string;
    demeanor_value: number;
    demeanor_category: string;
  }
  